Moved SearchParameterManager logic out of App.tsx
https://coveord.atlassian.net/browse/KIT-415
