--- packages/samples/headless-react/src/components/search-parameter-manager/search-parameter-serializer.ts
@@ -1,13 +1,0 @@
-import {buildSearchParameterSerializer} from '@coveo/headless';
-
-const {serialize, deserialize} = buildSearchParameterSerializer();
-
-export function readSearchParametersFromURI() {
-  return deserialize(decodeURIComponent(window.location.hash.slice(1)));
-}
-
-type SearchParameters = ReturnType<typeof readSearchParametersFromURI>;
-
-export function writeSearchParametersToURI(searchParameters: SearchParameters) {
-  window.location.hash = serialize(searchParameters);
-}
